requirements.txt
pandas==1.5.0
psycopg2==2.9.6
python-dotenv==1.0.0
